deploy: 2ffd746fd9c811ac7e0c845e5ed7554ea345a8f8

+++ apps/spiral/builder/build.ps1
@@ -1,6 +1,8 @@
 param(
     $fast,
     $SkipNotebook,
+    $SkipFsx,
+    $SkipPreBuild,
     $ScriptDir = $PSScriptRoot
 )
 Set-Location $ScriptDir
@@ -11,33 +13,44 @@ $ErrorActionPreference = "Stop"
 
 $projectName = "spiral_builder"
 
-if (!$fast -and !$SkipNotebook) {
-    { . ../dist/Supervisor$(_exe) --execute-command "../../../workspace/target/release/spiral_builder$(_exe) dib --path $projectName.dib" } | Invoke-Block -Retries 3
+if (!$SkipPreBuild -and !$SkipFsx) {
+    if (!$fast -and !$SkipNotebook) {
+        { . ../dist/Supervisor$(_exe) --execute-command "../../../workspace/target/release/spiral_builder$(_exe) dib --path $projectName.dib" } | Invoke-Block -Retries 3
+    }
+
+    { . ../../parser/dist/DibParser$(_exe) "$projectName.dib" spi } | Invoke-Block
+
+    { . ../dist/Supervisor$(_exe) --build-file "$projectName.spi" "$projectName.fsx" } | Invoke-Block
 }
 
-{ . ../../parser/dist/DibParser$(_exe) "$projectName.dib" spi } | Invoke-Block
+if (!$SkipPreBuild) {
+    $runtime = $fast -or $env:CI ? @("--runtime", ($IsWindows ? "win-x64" : "linux-x64")) : @()
+    $builderArgs = @("$projectName.fsx", "--persist-only", $runtime, "--packages", "Fable.Core", "--modules", @(GetFsxModules), "lib/fsharp/Common.fs")
+    { . ../../builder/dist/Builder$(_exe) @builderArgs } | Invoke-Block
 
-{ . ../dist/Supervisor$(_exe) --build-file "$projectName.spi" "$projectName.fsx" } | Invoke-Block
+    $targetDir = GetTargetDir $projectName
 
-$runtime = $fast -or $env:CI ? @("--runtime", ($IsWindows ? "win-x64" : "linux-x64")) : @()
-$builderArgs = @("$projectName.fsx", "--persist-only", $runtime, "--packages", "Fable.Core", "--modules", @(GetFsxModules), "lib/fsharp/Common.fs")
-{ . ../../builder/dist/Builder$(_exe) @builderArgs } | Invoke-Block
+    { BuildFable $targetDir $projectName "rs" } | Invoke-Block
+    (Get-Content "$targetDir/target/rs/$projectName.rs") `
+        -replace "../../../../lib", "../../../lib" `
+        -replace ".fsx`"]", ".rs`"]" `
+        | FixRust `
+        | Set-Content "$projectName.rs"
 
-$targetDir = GetTargetDir $projectName
+    if ($env:CI) {
+        Remove-Item $targetDir -Recurse -Force -ErrorAction Ignore
+    }
+}
 
-{ BuildFable $targetDir $projectName "rs" } | Invoke-Block
-(Get-Content "$targetDir/target/rs/$projectName.rs") `
-    -replace "../../../../lib", "../../../lib" `
-    -replace ".fsx`"]", ".rs`"]" `
-    | FixRust `
-    | FixRust2 `
+(Get-Content "$projectName.rs") `
+    -replace `
+        ($IsWindows ? "std::os::unix::fs::symlink" :"std::os::windows::fs::symlink_dir"), `
+        ($IsWindows ? "std::os::windows::fs::symlink_dir" :"std::os::unix::fs::symlink") `
     | Set-Content "$projectName.rs"
 
 cargo fmt --
 
-{ cargo +nightly test --release -- --show-output } | Invoke-Block
-{ cargo +nightly build --release } | Invoke-Block -OnError Continue
-
-if ($env:CI) {
-    Remove-Item $targetDir -Recurse -Force -ErrorAction Ignore
+if (!$fast) {
+    { cargo +nightly test --release -- --show-output } | Invoke-Block
 }
+{ cargo +nightly build --release } | Invoke-Block -OnError Continue
